# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "人事管理后端 API 文档（含 admin/user/chat 等接口）",
        "title": "Gin Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/captcha": {
            "get": {
                "description": "获取注册/登录验证码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取验证码",
                "parameters": [
                    {
                        "enum": [
                            "register",
                            "login"
                        ],
                        "type": "string",
                        "default": "register",
                        "description": "验证码场景",
                        "name": "scene",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CaptchaSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "账号登录，必要时需要验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "注册新账号（需验证码）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.APIErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "err": {
                    "type": "string",
                    "example": "error detail"
                },
                "error": {
                    "type": "string",
                    "example": "error detail"
                },
                "fail_count": {
                    "type": "integer",
                    "example": 3
                },
                "msg": {
                    "type": "string",
                    "example": "请求失败"
                },
                "need_captcha": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "handlers.CaptchaData": {
            "type": "object",
            "properties": {
                "captcha_id": {
                    "type": "string",
                    "example": "rY8M8P4KQd"
                },
                "expire_seconds": {
                    "type": "integer",
                    "example": 180
                },
                "image_base64": {
                    "type": "string",
                    "example": "data:image/png;base64,..."
                },
                "scene": {
                    "type": "string",
                    "example": "register"
                }
            }
        },
        "handlers.CaptchaSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/handlers.CaptchaData"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "handlers.LoginSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "type": "string",
                    "example": "登录成功"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "handlers.RegisterSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "type": "string",
                    "example": "注册成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handlers.SwaggerLoginRequest": {
            "type": "object",
            "properties": {
                "captcha_code": {
                    "type": "string",
                    "example": "A3D7"
                },
                "captcha_id": {
                    "type": "string",
                    "example": "rY8M8P4KQd"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "handlers.SwaggerRegisterRequest": {
            "type": "object",
            "properties": {
                "captcha_code": {
                    "type": "string",
                    "example": "A3D7"
                },
                "captcha_id": {
                    "type": "string",
                    "example": "rY8M8P4KQd"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "role": {
                    "type": "string",
                    "example": "staff"
                },
                "username": {
                    "type": "string",
                    "example": "staff01"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入格式: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}